+++ CHANGELOG.md
@@ -5,9 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [10.3.0] - 2021-06-09
 
+### Added
+
+- Added Console.size setter
+- Added Console.width setter
+- Added Console.height setter
+- Added angular style Rich reprs
 - Added an IPython extension. Load via `%load_ext rich`
+
+### Changed
+
+- Changed the logic for retrieving the calling frame in console logs to a faster one for the Python implementations that support it.
 
 ## [10.2.2] - 2021-05-19